+++ src/DistributionAlgo.h
@@ -20,16 +20,16 @@ class DistributionAlgo {};
 class DistributePoints : public DistributionAlgo
 {
 public:
-    DistributePoints(const NameValuePairs<Int> &p) : projections(p) {}
+	DistributePoints(const NameValuePairs<Int64> &p) : projections(p) {}
     
 protected:
-    NameValuePairs<Int> projections;
+	NameValuePairs<Int64> projections;
 };
 
 struct DistribuePointsAvg : DistributePoints
 {
-    DistribuePointsAvg(const NameValuePairs<Int> &p) : DistributePoints(p) {}
-    NameValuePairs<double> distribute(const Int result, const std::string &agent ) const;
+	DistribuePointsAvg(const NameValuePairs<Int64> &p) : DistributePoints(p) {}
+	NameValuePairs<double> distribute(const Int64 result, const std::string &agent) const;
 };
